Updated the login auth route
Updated the auth route to use the users model instead of the User class for login functionality.
Added the company to users and user to company associations.
Added last_login to users model.
Implemented last_login functionality in the auth login route.
Removed is_deleted from the users model as the table tracks deleted users with the paranoid setting.
Reverted testing changes made to the userController.